+++ source/extensions/lightspeed.trex.properties_pane.logic.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.3]
+### Changed
+- Updated to use `TreeItemBase.parent` setter for adding children (compatibility with omni.flux.utils.widget 1.24.0)
+
 ## [1.3.2]
 ### Changed
 - Switched to ruff for linting and formatting